Some improvements and code organization. Put PayPal operations in PayPal class.
