+++ src/main/java/io/github/proxyprint/kitchen/models/repositories/ManagerDAO.java
@@ -1,6 +1,7 @@
 package io.github.proxyprint.kitchen.models.repositories;
 
 import io.github.proxyprint.kitchen.models.printshops.Manager;
+import io.github.proxyprint.kitchen.models.printshops.PrintShop;
 import org.springframework.data.repository.CrudRepository;
 
 import javax.transaction.Transactional;
@@ -13,4 +14,6 @@ public interface ManagerDAO extends CrudRepository<Manager,Long> {
 
     @Transactional
     public Manager findByUsername(String username);
+    @Transactional
+    public Manager findByPrintShop(PrintShop printShop);
 }
